docs(swagger): rename docstring_parsing.py
